requirements.txt
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0
networkx>=2.6.0
plotly>=5.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
joblib>=1.1.0
pydantic>=1.8.0
click>=8.0.0
praw>=7.0.0
sentence-transformers>=2.0.0
transformers>=4.0.0
torch>=1.9.0
faiss-cpu>=1.7.0
pyarrow>=5.0.0
h5py>=3.3.0
numba>=0.56.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
python-dotenv>=0.19.0
redis>=4.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0